tests: move EnvSerializer and IniSerializer to tests/Common
